requirements.txt
numpy
scipy
pandas
matplotlib
jupyter
ipython
nuts-finder
setuptools
git+https://github.com/nestauk/data_getters.git
ratelim


# local package
-e .

# external requirements
dvc
jq
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
